CU012: PANTALLA DE ENTREGAR RESERVAS REALIZADA Y CONECTADA FUNCIONALIDADES TERMINADAS
