+++ src/main/resources/templates/pantallaAdministrador.html
@@ -170,7 +170,7 @@
     }
 
     function entregarReservas() {
-        mostrarModal('/entregarReservas.html');
+        mostrarModal('/reservas/pendientes');
     }
 
     function mostrarModal(url) {
